Add connection udp feedback

+++ DeusClientCore/DeusClientCore/Packets/General/PacketTextMessage.cs
@@ -17,13 +17,13 @@ namespace DeusClientCore.Packets
         public override ushort EstimateCurrentSerializedSize()
         {
             // nb we serialize the size of the string that's why we need 4 bytes before the string
-            return (ushort)(sizeof(uint) + (MessageText.Length + 1) );
+            return (ushort)(sizeof(UInt16) + (MessageText.Length + 1) );
         }
 
         public override void OnDeserialize(byte[] buffer, int index)
         {
             // get the size of the string
-            uint dataSize = 0;
+            UInt16 dataSize = 0;
             Serializer.DeserializeData(buffer, ref index, out dataSize);
 
             string message;
@@ -36,7 +36,7 @@ namespace DeusClientCore.Packets
             List<byte> result = new List<byte>();
 
             //uint dataSize = (uint)(MessageText.Length + 1); // +1 to add the \0 of string
-            result.AddRange(Serializer.SerializeData((uint)(MessageText.Length + 1)));
+            result.AddRange(Serializer.SerializeData((UInt16)(MessageText.Length + 1)));
 
             //  then we add the string 
             result.AddRange(Serializer.SerializeData(MessageText));
